MaxTwoRecordingExtractor
