--- spikeextractors/extractors/maxonerecordingextractor/__init__.py
@@ -1,1 +1,0 @@
-from .maxonerecordingextractor import MaxOneRecordingExtractor
